Add MMDB_aget_value function
This complement MMDB_get_value and MMDB_vget_value by allowing the caller to
pass a plain list of strings as the lookup path. Only allowing varargs or a
va_list makes certain use cases nearly impossible, since it's very difficult
to convert a list of strings to a va_list (as opposed to going the other way
around).

I also added a test for this.