Since we're going to read or mmap the entire file, just use that to find metadata
Previously we were reading the last 128kb of the file into a separate variable
and using that to find the metadata, but that was kind of pointless. We have
to read/mmap the whole file anyway.

I also simplified the metadata reading code a little while I was in there.