Avoid integer overflow on very large files.
This affects the non-Windows version of map_file, which is called
during MMDB_open:

File sizes are generally handled with off_t data type. The width
of this data type is flexible, but generally 64 bit. In map_file(),
the file size is converted from off_t to ssize_t, which is also a
flexible data type. On 32 bit systems, it's generally 32 bit.

Therefore, if the database file is larger than 4 GB, a truncation
can occur on such 32 bit systems.

To fix this, simply check if the value was truncated. I chose to
flag it as "out of memory", because you can argue that you are out
of memory for such a large file -- even though you will never have
enough memory with such an architecture to begin with.