Verify existence of metadata marker
The current code does not properly check if there is at least one
marker inside the database. If there is no marker available, the
initial search area will be considered as a valid metadata start,
which possibly leads to out of boundary access due to offset addition.

An example would be a database file with only one byte in it (1 byte is
needed for a successful mmap call). This statement would lead to
out of boundary access later on (file_size is 1):

    *metadata_start = search_area + strlen(METADATA_MARKER);

Also, if there are multiple markers in the metadata area, use the
last one encountered. The old code looks like that was intended.