Fix wrong epoch formatting for verbose mmdblookup (at least on 32 bit builds on Solaris Sparc).
Verbose mmdblookup output before the fix:
   Build epoch:   1497876636 (1970-01-01 00:00:00 UTC)
and after the fix:
   Build epoch:   1497876636 (2017-06-19 12:50:36 UTC)

The original code converted *uint64_t to *time_t by casting
the pointer. But that doesn't convert the referenced value
and for a build where a time_t is a long of size 32bits
(eg. on Solaris Sparc) probably also depending on endianess
when dereferencing the casted pointer you get the wrong half
of the 64 bits (=0).

The fix is to cast the value and use a pointer to the casted value.