Treat entry data list pointers as invalid when we grow the memory pool
The memory pool used to reduce allocations can move around. The calling
code using the pool held on to pointers in a few cases where it should
not have. This lead to both invalid reads and writes.

We now look up where the entry data list elements are each time the
pool's location may have moved prior to using them.