Check for O_CLOEXEC directly rather than POSIX version
POSIX 2008 provides this and when we request POSIX 2008 we will have it
exposed if the implementation provides it. We can check for it directly
rather than looking at what version the implementation provides. This is
more targeted, and should be the same thing (unless it's not implemented
as a macro I suppose, but that seems unlikely).