Add lowdown as an alternative Markdown->man translator
lowdown is a simple Markdown translator to multiple formats, including
man. lowdown is portable, and has no external dependencies, unlike
Pandoc which, due to its dependencies on the Haskell stack, can be a
headache for porters.

To make this more DRY, it converts the metadata preamble from Pandoc's
custom format ("% title(section") to Multimarkdown's, which is supported
by both Pandoc and lowdown.

Finally, this adds support for multiple translators in the
make-man-pages.pl script, to provide extensibility in case others (like
e.g. ronn-ng) are considered suitable in the future.

Fixes: #225