Disable "smarty" filters from manpage generation
From Pandoc's documentation on the "smart" extension:
  Interpret straight quotes as curly quotes, --- as em-dashes, -- as
  en-dashes, and ... as ellipses. Nonbreaking spaces are inserted after
  certain abbreviations, such as “Mr.”

This has the effect that all command-line options, like "--file", are
translated as "\[en]file". That is, an en-dash, looking like a larger
hyphen and potentially confusing users.

Rather than escape these in the Markdown source, pass options to both
pandoc and lowdown to disable smart(y) features.

Note that this has the side-effect of disabling some curly quotes as
well, but these are missing from the GitHub HTML output as well.