make-man-pages: small tidying
Rename arguments and expand system() calls into multiple lines, to match
what geoipupdate's make-man-pages.pl does.

This will become handy in the next commit, as we add more arguments to
pandoc/lowdown, as well as in future convergence attempts with
geoipupdate.

While at it, also pass the code through perltidy, which changes
whitespace in a couple of other places.